feat: Add Twitter handler for X Bot endpoints
